pyproject.toml
[project]
name = "chronos"
version = "0.1.0"
description = "AI Agentic Planner"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.0.14",
    "groq>=0.9.0",
    "aiogram>=3.3.0",
    "sqlmodel>=0.0.14",
    "aiosqlite>=0.20.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.0",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
